paper-trading/requirements.txt
ccxt>=4.0.0
flask>=3.0.0
pandas>=2.0.0
ta>=0.11.0
plotly>=5.18.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
